# Move DESCRIBE CONFIG into the compiler