Currently it gets precompiled at bootstrap time and cached in
the functions _describe_system_config_as_ddl and
_describe_database_config_as_ddl

Once we have extension/user-creatable configs, having a single
pre-compiled config description function won't work.

Here, I've fixed it by moving it out of the function and recompiling
it in the compiler at query execution time. This is obviously slower
but I don't think this is on critical paths?

Another option would be to regenerate _describe_database_config_as_ddl
whenever a config object is created/deleted/altered.

I think that would work, and would have better performance
characteristics than this, but I'm not totally comfortable with having
standard library functions dynamically getting recompiled based on the
contents of the user schema. Maybe I should just get over that, though?