@elprans 2023-08-21
How much slower is this compared to the current implementation?