@msullivan 2023-08-21
Current version: 250ms uncached, 215ms cached
New version: 770ms uncached, 20ms cached

Looks like the actual query performance is quite a bit better with the non-function version!
(At the cost of our compiler, of course)